refactor axiosInterceptor as function
